Move "Mark this page as read" to the bottom
This better matches the most common use case for the "Mark this page as
read" link, which is to click it once you have reached the bottom of the
page. Especially on mobile, it can be annoying to scroll to the top of
the page to click the link.

On the unread entries page, the "Mark all as read" link is still at the
top because it is unlikely to be used in the same way as the other link.

Closes #196 (https://github.com/miniflux/miniflux/issues/196).