Add Server-Timing header to unread page
To be used as a basis to discuss #722. Current metrics may not be the
most interesting ones.

This is not deactivable at the moment, the header is:
```
server-timing: prep;dur=30.548655,sql_unread;dur=8.762088,sql_entries;dur=16.670605,render;dur=566.738683
```

The construction advised by the library
```
defer timing.NewMetric(name).Start().Stop()
```
does not work, the metric has a duration of 0s. I suspect this is due to
the request being finalized too early for the header to be added. Some
more granular metrics as added in this patch may be more relevant anyway.