feed_entries: Add mark feed as read option
Why:
e1c9e6ccb4607eeab2 added the support to mark a feed as read from views
showing a list of feeds. However, at times it can be useful to mark an
entire feed as read from within the feed entry views, e.g. when wanting
to mark as read a feed that is dominating a category after examining a
number of entries in that feed

What:
Add an li element in feed_entries.html to allow marking the feed as read
The code has been copy pasted from e1c9e6ccb4607eeab2

This closes #721