Redirect to the view the request came from when marking feed as read
Piggyback on the javascript used for POSTing to a url to obtain a more
consistent behavior with the other options in the feed list (e.g. remove
feed). That way, when a feed is marked as read, regardless which view
the request originated from (feeds list view vs category feeds list
view), we are redirected to the view we started from.

While at it, add a confirmation dialog to mark feed as read. It was
suggested in PR #1055 for the single feed option, but it probably makes
sense if it was in the feeds list view too.