Remove swiping on feed links on mobile
The swiping does not work very well on all recent mobiles (multiple
iPhones and several Android tested). Most often, the touch event is
triggered when trying to scroll.  This makes the scrolling on mobile
a not so pleasant experience. When swiping, it oftens scrolls
instead.

Since the buttons at the botton of each link can do similarly without
being in the way, this PR proposes to keep only thoses and to remove
the touch event on feed links.