fix for wrong feed entry relation
In Reeder some entries appeared to have been incorrectly associated with an incorrect feed.
This seems to have been a problem with the /reader/api/0/stream/items/contents api endpoint.
This endpoint contains two occurences of the feeds id in the form "feed/$DECIMAL_NUMBER".
The decimal number was encoded in hex, which seems to have caused the wrong association.