Add individual passkey management
Introduce a list of passkeys and options to remove and rename them.
Because the native library data structure does not have space for this
new metadata, this change introduces a wrapping struct to old on to that
data and hence there are quite a few changes to interfaces.

It is also quite tricky to identify passkeys just based on the
information they give you, so inspired by Github, you are now prompted
to create a name for your passkey. Combined with when they were created
and last used that should make management easier. The field is
intentionally optional if users don't want to enter a name. When
registering one they will now be prompted, but can choose to ignore that
or enter an empty name. Registration has already succeeded at that
point.