calling store.UpdateFeed and store.UpdateFeedError in a defer function
This makes sure that we are always calling the following functions:
* originalFeed.ScheduleNextCheck
* store.UpdateFeed
* store.UpdateFeedError
And we only call them in a single place regardless which branch is taken inside function.

The store functions within them will udpate the value in database.

This change aims to reduce "unknown unknowns" in the codes. With this change, we won't miss calling above functions.