refactor(js): golf a bit touch-related js
There is no need to instantiate the whole TouchHandler object to call a single
static method that immediately returns in case touch events aren't supported:
it's better to extract the condition around the calls, so that the object is
never instantiated.

Another change is to replace `() => this.reset()` with `this.reset`, which is
equivalent, but shorted and less weird.