refactor(sanitizer): use a smarter approach wrt. SanitizeHTML
The issue with slices.Contains is that it keeps growing and never shrinks. An
rss entry could start with a bunch of random tags, and those would always be
considered valid in the `case html.EndTagToken` case.

While can't use a strict push/pop queue, as we might have things like
<video><source></video> that are valid HTML, we can look in the slice for the
farther index of the tag that is being closed, and pop everything after it.