refactor(config): massively simplify the configuration parser
Instead of having as structure with every possible type as member, use a single
`any`. The type is determined by the default value, reducing the amount of
copy/paste.

This PR is sent as a draft to gather feedback before moving all the different
types to this new parsing method.